@@ -20,8 +20,8 @@ export default {
   },
   computed: {
     results () {
-      let results = []
-      for (let item in this.items) {
+      const results = []
+      for (const item in this.items) {
         if (this.items[item].title.search(new RegExp(this.text, 'i')) > -1 || this.items[item].description.search(new RegExp(this.text, 'i')) > -1) {
           results.unshift(this.items[item])
           results[0].id = item
